requirements.txt
python-chess>=1.999
mcp>=1.0
fastapi>=0.115
uvicorn>=0.30
openai>=1.50
python-dotenv>=1.0
pandas
numpy
